% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exclude_lengths}
\alias{exclude_lengths}
\title{Exclude length measurements}
\usage{
exclude_lengths(metrics, bad_metrics, metadata)
}
\arguments{
\item{metrics}{A tibble of metric data using standard abbreviations
(after von den Driesch 1976)}

\item{metadata}{List of non-metric column names}
}
\value{
A filtered tibble
}
\description{
Filter length measurements, that aren't correlated with body mass, from metric
data (see Meadows 1999), except for astragalus GLI.
}
