% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lsi}
\alias{lsi}
\title{Log Standard Index}
\usage{
lsi(obs, std, obs_metadata, std_metadata)
}
\arguments{
\item{obs}{Tibble of observed animal measurements}

\item{std}{Tibble of standard animal measurements}

\item{check}{TRUE (default) will print a warning when a measurement in obs is
not matched in std}
}
\value{
A tibble of log standard indexes.
}
\description{
Calculates the log standard index (LSI) or difference-of-logs index, which
standardises measurements from different skeletal elements with reference
to a "standard animal". Specifically:
}
\details{
LSI = log X - log S

Where X is the measurement of the specimen and S is the same measurement of
the same element from a standard reference animal.
}
